+++ src/sys/lib/rt.vdso/src/rt_process.rs
@@ -969,7 +969,7 @@ fn encode_env(keys: Vec<String>, vals: Vec<String>) -> Result<u64, ErrorCode> {
     }
 
     unsafe {
-        let mut pos = result_addr as usize;
+        let mut pos = result_addr;
         *((pos as *mut u32).as_mut().unwrap()) = num_args;
         pos += 4;
